+++ ChangeLog.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Changed
+- Support for Draco mesh compression is now optional (install DracoUnity package to enable it)
+
 ## [0.11.0] - 2020-03-07
 ### Added
 - Support for texture samplers' wrapping mode